% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brm_formula.R
\name{brm_formula}
\alias{brm_formula}
\title{Model formula}
\usage{
brm_formula(
  response = "CHG",
  group = "TRT01P",
  time = "AVISIT",
  base = "BASE",
  covariates = character(0),
  intercept = TRUE,
  effect_group = TRUE,
  effect_time = TRUE,
  effect_base = TRUE,
  interaction_base = TRUE,
  interaction_group = TRUE
)
}
\arguments{
\item{response}{Character of length 1, name of the response variable.}

\item{group}{Character of length 1, name of the treatment group
variable.}

\item{time}{Character of length 1, name of the discrete time variable.}

\item{base}{Character of length 1, name of the baseline response variable.}

\item{covariates}{Character vector of names of other covariates.}

\item{intercept}{\code{TRUE} to include an intercept, \code{FALSE} to omit.}

\item{effect_group}{\code{TRUE} to include an additive effect for treatment
group, \code{FALSE} to omit.}

\item{effect_time}{\code{TRUE} to include an additive effect for discrete
time, \code{FALSE} to omit.}

\item{interaction_base}{\code{TRUE} to include baseline-by-time interaction,
\code{FALSE} to omit.}

\item{interaction_group}{\code{TRUE} to include treatment-group-by-time
interaction, \code{FALSE} to omit.}
}
\value{
A formula to specify an MMRM.
}
\description{
Build a model formula for an MMRM.
}
\examples{
brm_formula()
brm_formula(intercept = FALSE, effect_base = FALSE)
brm_formula(
  intercept = FALSE,
  effect_base = FALSE,
  interaction_group = FALSE
)
}
\seealso{
Other model: 
\code{\link{brm_model}()}
}
\concept{model}
