% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brm_model.R
\name{brm_model}
\alias{brm_model}
\title{Basic MMRM}
\usage{
brm_model(
  data,
  formula = brm_formula(),
  correlation = "unstructured",
  time = "AVISIT",
  patient = "USUBJID",
  sd_intercept = 100,
  sd_b = 100,
  sd_intercept_sigma = 100,
  sd_b_sigma = 100,
  ...
)
}
\arguments{
\item{data}{A tidy data frame with one row per patient per discrete
time point.}

\item{formula}{An R formula to represent the fixed effect parameterization.
Use \code{\link[=brm_formula]{brm_formula()}} to help.}

\item{correlation}{Character of length 1, name of the correlation
structure. Only \code{"unstructured"} is currently supported.}

\item{sd_intercept}{Positive numeric of length 1, prior standard deviation
of the "intercept" class of parameters.}

\item{sd_b}{Positive numeric of length 1, prior standard deviation
of the "b" class of parameters.}

\item{sd_intercept_sigma}{Positive numeric of length 1,
prior standard deviation
of the "intercept" class of parameters with \code{dpar = "sigma"}.}

\item{sd_b_sigma}{Positive numeric of length 1,
prior standard deviation
of the "b" class of parameters with \code{dpar = "sigma"}.}

\item{...}{Arguments to \code{brms::brm()} other than \code{data}, \code{formula},
and \code{prior}.}
}
\value{
A fitted model object from \code{brms}.
}
\description{
Fit a basic MMRM model using \code{brms}.
}
\examples{
set.seed(0L)
data <- brm_simulate()$data
formula <- brm_formula(
  response = "response",
  group = "group",
  time = "time",
  effect_base = FALSE,
  interaction_base = FALSE
)
tmp <- utils::capture.output(
  suppressMessages(
    suppressWarnings(
      model <- brm_model(
        data = data,
        formula = formula,
        time = "time",
        patient = "patient",
        chains = 1,
        iter = 100,
        refresh = 0
      )
    )
  )
)
model
brms::prior_summary(model)
}
\seealso{
Other model: 
\code{\link{brm_formula}()}
}
\concept{model}
