% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brm_plot_compare.R
\name{brm_plot_compare}
\alias{brm_plot_compare}
\title{Visualize the model against the data.}
\usage{
brm_plot_compare(summaries_draws, summaries_data)
}
\arguments{
\item{summaries_draws}{Posterior summaries of marginal MCMC draws
from \code{\link[=brm_marginal_summaries]{brm_marginal_summaries()}}.}

\item{summaries_data}{Posterior summaries of the data
from \code{\link[=brm_marginal_data]{brm_marginal_data()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualize the model against the data.
}
\examples{
set.seed(0L)
sim <- brm_simulate()
data <- sim$data
data$group <- paste("treatment", data$group)
data$time <- paste("visit", data$time)
formula <- brm_formula(
  response = "response",
  group = "group",
  time = "time",
  patient = "patient",
  effect_base = FALSE,
  interaction_base = FALSE
)
tmp <- utils::capture.output(
  suppressMessages(
    suppressWarnings(
      model <- brm_model(
        data = data,
        formula = formula,
        chains = 1,
        iter = 100,
        refresh = 0
      )
    )
  )
)
draws <- brm_marginal_draws(
  model = model,
  group = "group",
  time = "time",
  patient = "patient",
  control = "treatment 1",
  baseline = "visit 1",
  outcome = "response"
)
suppressWarnings(summaries_draws <- brm_marginal_summaries(draws))
summaries_data <- brm_marginal_data(
  data,
  response = "response",
  group = "group",
  time = "time"
)
brm_plot_compare(summaries_draws, summaries_data)
}
\concept{visualization}
