% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brm_prior.R
\name{brm_prior_label}
\alias{brm_prior_label}
\title{Label a prior with levels in the data.}
\usage{
brm_prior_label(
  label = NULL,
  code = NA_character_,
  group = NA,
  subgroup = NA,
  time = NA
)
}
\arguments{
\item{label}{A \code{tibble} with the prior labeling scheme so far,
with one row per model parameter and columns for the
Stan code, treatment group, subgroup, and discrete time point
of each parameter.}

\item{code}{Character of length 1, Stan code for the prior. Could be
a string like \code{"normal(1, 2.2)"}. The full set of priors is given
in the Stan Reference Manual, available from \url{https://mc-stan.org/}.
See the documentation \code{\link[brms:set_prior]{brms::set_prior()}} for more details.}

\item{group}{Value of length 1, level of the treatment group column
in the data to label the prior. The treatment group column
is the one you identified with the \code{group} argument of \code{\link[=brm_data]{brm_data()}}.}

\item{subgroup}{Value of length 1, level of the subgroiup column
in the data to label the prior. The subgroup column
is the one you identified with the \code{subgroup} argument of \code{\link[=brm_data]{brm_data()}}.
Not every dataset has a subgroup. If yours does not, please use
the default value \code{NA}.}

\item{time}{Value of length 1, level of the discrete time column
in the data to label the prior. The discrete time column
is the one you identified with the \code{time} argument of \code{\link[=brm_data]{brm_data()}}.}
}
\value{
A \code{tibble} with one row per model parameter and columns for the
Stan code, treatment group, subgroup, and discrete time point
of each parameter.
}
\description{
Label an informative prior for a parameter
using a collection of levels in the data.
}
\details{
\code{\link[=brm_prior_archetype]{brm_prior_archetype()}} builds a \code{brms} prior for an
informative prior archetype model. Successive calls to
\code{\link[=brm_prior_parameter]{brm_prior_parameter()}} build the prior specificiation for the \code{label}
argument of \code{\link[=brm_prior_archetype]{brm_prior_archetype()}}.
}
\section{Prior labeling}{

Informative prior archetypes use a labeling scheme to assign
priors. Priors are labeled using levels in the data, and then
\code{\link[=brm_prior_archetype]{brm_prior_archetype()}} uses the labels to map each prior to its
corresponding parameter in the model. This makes it possible
to assign informative priors without having to guess the names
of specific model coefficients in \code{brms}. The specific mapping is
varied from archetype to archetype, and it is explained in
specific archetype functions such as \code{\link[=brm_archetype_successive_cells]{brm_archetype_successive_cells()}}.
}

\examples{
set.seed(0L)
data <- brm_simulate_outline(
  n_group = 2,
  n_patient = 100,
  n_time = 3,
  rate_dropout = 0,
  rate_lapse = 0
) |>
  dplyr::mutate(response = rnorm(n = dplyr::n())) |>
  brm_simulate_continuous(names = c("biomarker1", "biomarker2")) |>
  brm_simulate_categorical(
    names = c("status1", "status2"),
    levels = c("present", "absent")
  )
archetype <- brm_archetype_successive_cells(data)
dplyr::distinct(data, group, time)
label <- brm_prior_label(
  code = "normal(1, 1)",
  group = "group_1",
  time = "time_1"
) |>
  brm_prior_label("normal(1, 2)", group = "group_1", time = "time_2") |>
  brm_prior_label("normal(1, 3)", group = "group_1", time = "time_3") |>
  brm_prior_label("normal(2, 1)", group = "group_2", time = "time_1") |>
  brm_prior_label("normal(2, 2)", group = "group_2", time = "time_2") |>
  brm_prior_label("normal(2, 3)", group = "group_2", time = "time_3")
label
}
\seealso{
Other priors: 
\code{\link{brm_prior_simple}()}
}
\concept{priors}
