% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brm_scenario_successive.R
\name{brm_scenario_successive}
\alias{brm_scenario_successive}
\title{Successive differences scenario}
\usage{
brm_scenario_successive(data)
}
\arguments{
\item{data}{A special classed \code{tibble} returned from \code{\link[=brm_data]{brm_data()}}.
It should have one row per patient per discrete time point.}
}
\value{
A special classed \code{tibble} with pre-processed data tailored to
the successive differences scenario.
}
\description{
Create an informative prior scenario where the fixed effects
are successive differences between adjacent time points.
}
\details{
In this parameterization, each fixed effect is either an intercept
on the first time point or the difference between two adjacent time
points, and each treatment group has its own set of fixed effects
independent of the other treatment groups.

To illustrate, suppose the dataset has two treatment groups A and B,
time points 1, 2, and 3, and no other covariates.
Let \code{mu_gt} be the conditional expectation of the response at group
\code{g} time \code{t} given data and hyperparameters. Then, the cell-means-like
successive differences parameterization declares model coefficients
\code{beta_1}, \code{beta_2}, ... \code{beta_6} as follows:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  `mu_A1 = beta_1`
  `mu_A2 = beta_1 + beta_2`
  `mu_A3 = beta_1 + beta_2 + beta_3`

  `mu_B1 = beta_4`
  `mu_B2 = beta_4 + beta_5`
  `mu_B3 = beta_4 + beta_5 + beta_6`
}\if{html}{\out{</div>}}

For group A, \code{beta_1} is the time 1 intercept, \code{beta_2} represents
time 2 minus time 1, and \code{beta_3} represents time 3 minus time 2.
\code{beta_4}, \code{beta_5}, and \code{beta_6} represent the analogous roles.
}
\section{Nuisance variables in informative prior scenarios}{

In the presence of other covariates, functions like
\code{\link[=brm_scenario_successive]{brm_scenario_successive()}} convert nuisance factors into binary
dummy variables, then center all those dummy variables and any
continuous nuisance variables at their means in the data.
This ensures that the main model coefficients
of interest are not implicitly conditional on a subset of the data.
In other words, preprocessing nuisance variables this way preserves
the interpretations of the fixed effects of interest, and it ensures
informative priors can be specified correctly.
}

\examples{
if (identical(Sys.getenv("BRM_EXAMPLES", unset = ""), "true")) {
set.seed(0L)
data <- brm_data(
  data = brm_simulate_simple()$data,
  outcome = "response",
  role = "response",
  group = "group",
  time = "time",
  patient = "patient",
  reference_group = "group_1",
  reference_time = "time_1"
)
data
brm_scenario_successive(data)
}
}
\concept{informative prior scenarios}
