% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brm_simulate_dropout.R
\name{brm_simulate_dropout}
\alias{brm_simulate_dropout}
\title{Simulate dropouts.}
\usage{
brm_simulate_dropout(data, rate)
}
\arguments{
\item{data}{A \code{tibble} generated by \code{\link[=brm_simulate_start]{brm_simulate_start()}}.}

\item{rate}{Positive numeric of length 1 between 0 and 1.
Expected dropout proportion observed by the final time point. Earlier time
points may have lower observed dropout proportions.}
}
\value{
The \code{tibble} from the \code{data} argument, but with a new or updated
column called \code{missing} which reflects the missingness pattern
of \code{\link[=brm_simulate_dropout]{brm_simulate_dropout()}}.
}
\description{
Simulate dropouts in a dataset initialized with
\code{\link[=brm_simulate_start]{brm_simulate_start()}}.
}
\details{
Initialize the simulation with \code{\link[=brm_simulate_start]{brm_simulate_start()}}, then
use \code{\link[=brm_simulate_dropout]{brm_simulate_dropout()}} to simulate a simple missingness pattern that
introduces a simple dropout pattern in the data.
Can be used in conjunction with other simulation functions which accept
a \code{data} argument, such as \code{\link[=brm_simulate_lapse]{brm_simulate_lapse()}}.

A dropout is an intercurrent event when data
collection for a patient stops permanently,
causing the outcomes for that patient to be missing during and after
the dropout occurred. The first time point is assumed to be baseline,
so dropout is there. Dropouts are equally likely to occur at each of
the post-baseline time points.
}
\examples{
library(magrittr)
brm_simulate_start() \%>\%
  brm_simulate_dropout()
}
\seealso{
Other simulation: 
\code{\link{brm_simulate_lapse}()},
\code{\link{brm_simulate_simple}()},
\code{\link{brm_simulate_start}()}
}
\concept{simulation}
