% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brm_simulate_lapse.R
\name{brm_simulate_lapse}
\alias{brm_simulate_lapse}
\title{Simulate lapsed outcomes.}
\usage{
brm_simulate_lapse(data, rate = 0.05)
}
\arguments{
\item{data}{A \code{tibble} generated by \code{\link[=brm_simulate_start]{brm_simulate_start()}}.}

\item{rate}{Positive numeric of length 1 between 0 and 1.
Expected proportion of missing values in the response due to
independent uniform sporadic lapses in data collection.}
}
\value{
The \code{tibble} from the \code{data} argument, but with a new or updated
column called \code{missing} which reflects the missingness pattern
of \code{\link[=brm_simulate_lapse]{brm_simulate_lapse()}}.
}
\description{
Simulate lapsed outcomes in a dataset initialized with
\code{\link[=brm_simulate_start]{brm_simulate_start()}}.
}
\details{
Initialize the simulation with \code{\link[=brm_simulate_start]{brm_simulate_start()}}, then
use \code{\link[=brm_simulate_lapse]{brm_simulate_lapse()}} to simulate a simple missingness pattern that
introduces independent uniform sporadic lapses in data collection.
Can be used in conjunction with other simulation functions which accept
a \code{data} argument, such as \code{\link[=brm_simulate_lapse]{brm_simulate_lapse()}}.
The earliest time point is treated as baseline and has no missing values.
}
\examples{
library(magrittr)
brm_simulate_start() \%>\%
  brm_simulate_lapse()
}
\seealso{
Other simulation: 
\code{\link{brm_simulate_dropout}()},
\code{\link{brm_simulate_simple}()},
\code{\link{brm_simulate_start}()}
}
\concept{simulation}
