% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brm_simulate_outline.R
\name{brm_simulate_outline}
\alias{brm_simulate_outline}
\title{Simulate a dataset outline.}
\usage{
brm_simulate_outline(
  n_group = 2L,
  n_patient = 100L,
  n_time = 4L,
  baseline = FALSE,
  rate_dropout = 0.1,
  rate_lapse = 0.025,
  ...
)
}
\arguments{
\item{n_group}{Positive integer of length 1, number of treatment groups.}

\item{n_patient}{Positive integer of length 1, number of patients
per treatment group.}

\item{n_time}{Positive integer of length 1, number of discrete
time points (e.g. scheduled study visits) per patient.}

\item{baseline}{Logical of length 1, whether the earliest time point
should be considered baseline. If \code{TRUE}, then the earliest time point
has no missing responses indicated in the \code{miss_dropout} or \code{miss_lapse}
columns of the output. Defaults to \code{FALSE} because \code{brms.mmrm} assumes
the outcome variable is change from baseline.}

\item{rate_dropout}{Positive numeric from 0 to 1,
expected proportion of observed dropouts at the final time point.
A dropout at time point \code{T} is an intercurrent event in which a patient
has missing responses at \code{T} and all time points after \code{T}.
For \code{\link[=brm_simulate_outline]{brm_simulate_outline()}}, the conditional probability of dropout
at time \code{T}, given that dropout did not occur at time \code{T} - 1, is the same
for all time points \code{T}.
The expected dropout rate is the same for each group.
The resulting missingness pattern is in the
\code{miss_dropout} column of the output:
\code{TRUE} if missing, \code{FALSE} if observed.}

\item{rate_lapse}{Positive numeric from 0 to 1,
expected proportion of observations
that are missing uniformly at random.
This source of missingness covers isolated temporary lapses
in data collection.
The resulting missingness pattern is in the
\code{miss_lapse} column of the output:
\code{TRUE} if missing, \code{FALSE} if observed.}
}
\value{
A data frame from \code{\link[=brm_data]{brm_data()}} with attributes to define roles
for various columns in the dataset.
}
\description{
Simulate an outline of an MMRM dataset.
}
\details{
This function simulates an outline of an MMRM dataset,
including covariates and a missingness pattern for the response.
However, it does not include the response variable. Simulating the
response from the prior predictive distribution requires a
fixed effect parameterization and a prior distribution, both of which
are outside the scope of \code{\link[=brm_simulate_outline]{brm_simulate_outline()}}.
}
\examples{
set.seed(0L)
brm_simulate_outline()
}
\seealso{
Other simulation: 
\code{\link{brm_simulate_simple}()}
}
\concept{simulation}
