% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brm_simulate_start.R
\name{brm_simulate_start}
\alias{brm_simulate_start}
\title{Start a simulated dataset}
\usage{
brm_simulate_start(n_group = 2L, n_patient = 100L, n_time = 4L)
}
\arguments{
\item{n_group}{Positive integer of length 1, number of treatment groups.}

\item{n_patient}{Positive integer of length 1, number of patients
per treatment group.}

\item{n_time}{Positive integer of length 1, number of discrete
time points (e.g. scheduled study visits) per patient.}
}
\value{
A data frame from \code{\link[=brm_data]{brm_data()}} with attributes to define roles
for various columns in the dataset. The data frame has one row per
patient per time point and the following columns:
\itemize{
\item \code{group}: integer index of the treatment group.
\item \code{patient}: integer index of the patient.
\item \code{time}: integer index of the discrete time point.
}
}
\description{
Begin creating a simulated dataset.
}
\examples{
brm_simulate_start()
}
\seealso{
Other simulation: 
\code{\link{brm_simulate_dropout}()},
\code{\link{brm_simulate_lapse}()},
\code{\link{brm_simulate_simple}()}
}
\concept{simulation}
