% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brm_summary.R
\name{brm_summary}
\alias{brm_summary}
\title{Summarize an MMRM.}
\usage{
brm_summary(
  model,
  base = "BASE",
  group = "TRT01P",
  time = "AVISIT",
  patient = "USUBJID",
  covariates = character(0),
  outcome = "change",
  control = "Placebo",
  baseline = "Baseline",
  level = 0.95,
  direction = "greater",
  threshold = 0
)
}
\arguments{
\item{model}{Fitted \code{brms} model object from \code{\link[=brm_model]{brm_model()}}.}

\item{base}{Character of length 1, name of the baseline response variable
in the data.}

\item{group}{Character of length 1, name of the treatment group
variable in the data.}

\item{time}{Character of length 1, name of the discrete time variable
in the data.}

\item{patient}{Character of length 1, name of the patient ID variable
in the data.}

\item{covariates}{Character vector of names of other covariates
in the data.}

\item{outcome}{Character of length 1, \code{"response"} if the
response variable is the raw outcome variable (such as AVAL)
or \code{"change"} if the response variable is change from baseline
(e.g. CHG).}

\item{control}{Element of the \code{group} column in the data which indicates
the control group for the purposes of calculating treatment differences.}

\item{baseline}{Element of the \code{time} column in the data
which indicates the baseline time for the purposes of calculating
change from baseline.}

\item{level}{Numeric of length 1 between 0 and 1, credible level
for the credible intervals. Only relevant when \code{return = "marginals"}.}

\item{direction}{Character vector of the same length as \code{threshold}.
\code{"greater"} to compute the marginal posterior probability that the
treatment effect is greater than the threshold,
\code{"less"} to compute the marginal posterior probability that the
treatment effect is less than the threshold.
Each element \code{direction[i]} corresponds to \code{threshold[i]}
for all \code{i} from 1 to \code{length(direction)}.}

\item{threshold}{Numeric vector of the same length as \code{direction},
treatment effect threshold for computing posterior probabilities.
Each element \code{direction[i]} corresponds to \code{threshold[i]} for
all \code{i} from 1 to \code{length(direction)}.}
}
\value{
A list of two data frames, one with summary statistics on the
marginal posterior and another with posterior probabilities
on the treatment effects.
}
\description{
Summarize a basic MMRM model fit.
}
\details{
Currently assumes the response variable is \code{CHG}
(change from baseline) and not \code{AVAL} (raw response).
}
\examples{
set.seed(0L)
sim <- brm_simulate()
data <- sim$data
data$group <- paste("treatment", data$group)
data$time <- paste("visit", data$time)
formula <- brm_formula(
  response = "response",
  group = "group",
  time = "time",
  patient = "patient",
  effect_base = FALSE,
  interaction_base = FALSE
)
tmp <- utils::capture.output(
  suppressMessages(
    suppressWarnings(
      model <- brm_model(
        data = data,
        formula = formula,
        chains = 1,
        iter = 100,
        refresh = 0
      )
    )
  )
)
brm_summary(
  model = model,
  group = "group",
  time = "time",
  patient = "patient",
  control = "treatment 1",
  baseline = "visit 1",
  outcome = "response"
)
}
\concept{results}
