% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brm_transform_marginal.R
\name{brm_transform_marginal}
\alias{brm_transform_marginal}
\title{Marginal mean transformation}
\usage{
brm_transform_marginal(
  data,
  formula,
  prefix = "b_",
  average_within_subgroup = NULL
)
}
\arguments{
\item{data}{Classed tibble with preprocessed data from \code{\link[=brm_data]{brm_data()}}.}

\item{formula}{Model formula from \code{\link[=brm_formula]{brm_formula()}}.}

\item{prefix}{Character of length 1, prefix to add to
the model matrix (\code{"X"}) from \code{\link[brms:standata]{brms::make_standata()}} in order to
reconstruct the \code{brms} model parameter names. This argument should
only be modified for testing purposes.}
}
\value{
A matrix to transform model parameters (columns) into
marginal means (rows).
}
\description{
Transformation from model parameters to marginal means.
}
\details{
The matrix from \code{\link[=brm_transform_marginal]{brm_transform_marginal()}} is passed to
the \code{transform_marginal} argument of \code{\link[=brm_marginal_draws]{brm_marginal_draws()}},
and it transforms posterior draws of model parameters to
posterior draws of marginal means. You may customize the output of
\code{\link[=brm_transform_marginal]{brm_transform_marginal()}} before passing it to \code{\link[=brm_marginal_draws]{brm_marginal_draws()}}.
However, please do not modify the dimensions, row names, or column
names.
}
\examples{
if (identical(Sys.getenv("BRM_EXAMPLES", unset = ""), "true")) {
set.seed(0L)
data <- brm_data(
  data = brm_simulate_simple()$data,
  outcome = "response",
  role = "response",
  group = "group",
  time = "time",
  patient = "patient",
  reference_group = "group_1",
  reference_time = "time_1"
)
formula <- brm_formula(
  data = data,
  baseline = FALSE,
  baseline_time = FALSE
)
transform <- brm_transform_marginal(data = data, formula = formula)
}
}
\concept{transformations}
