% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brm_package.R
\docType{package}
\name{brms.mmrm-package}
\alias{brms.mmrm-package}
\title{brms.mmrm: Bayesian MMRMs using \code{brms}}
\description{
The mixed model for repeated measures (MMRM) is a
popular model for longitudinal clinical trial data with
continuous endpoints, and \code{brms} a is powerful and versatile
package for fitting Bayesian regression models.
The \code{brms.mmrm} R package leverages \code{brms} to run MMRMs, and
it supports a simplified interfaced to reduce difficulty
and align with the best practices of the life sciences.
}
\references{
\itemize{
\item Paul-Christian Bürkner (2017). brms: An R Package for Bayesian
Multilevel Models Using Stan.
.    Journal of Statistical Software, 80(1), 1-28.
doi:10.18637/jss.v080.i01
\item Mallinckrodt, C.H., Lane, P.W., Schnell, D. et al.
Recommendations for the Primary Analysis of Continuous Endpoints
in Longitudinal Clinical Trials.
Ther Innov Regul Sci 42, 303–319 (2008).
doi:10.1177/009286150804200402
}
}
\concept{help}
