% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brm_archetype.R
\name{summary.brms_mmrm_archetype}
\alias{summary.brms_mmrm_archetype}
\title{Summarize an informative prior archetype.}
\usage{
\method{summary}{brms_mmrm_archetype}(object, message = TRUE, ...)
}
\arguments{
\item{object}{The informative prior archetype to summarize.}

\item{message}{TRUE to print an informative message about the archetype
and invisibly return a character vector of equations. \code{FALSE}
to forgo verbose messages and non-invisibly return the equations.}

\item{...}{Not used, but required for S3 methods that inherit from
the base generic \code{\link[=summary]{summary()}}.}
}
\value{
Return a character vector with linear equations
that map model parameters to marginal means. If the \code{message}
argument is \code{TRUE} (default) then this character vector is returned
invisibly and a verbose description of the equations is printed.
}
\description{
For an informative prior archetype, show
the transformation from model parameters to marginal means.
}
