% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brm_formula.R
\name{brm_formula}
\alias{brm_formula}
\title{Model formula}
\usage{
brm_formula(
  data,
  intercept = TRUE,
  effect_baseline = TRUE,
  effect_group = TRUE,
  effect_time = TRUE,
  interaction_baseline = TRUE,
  interaction_group = TRUE,
  correlation = "unstructured"
)
}
\arguments{
\item{data}{A classed data frame from \code{\link[=brm_data]{brm_data()}}.}

\item{intercept}{\code{TRUE} to include an intercept, \code{FALSE} to omit.}

\item{effect_baseline}{\code{TRUE} to include an additive effect for baseline
response, \code{FALSE} to omit.}

\item{effect_group}{\code{TRUE} to include an additive effects for treatment
groups, \code{FALSE} to omit.}

\item{effect_time}{\code{TRUE} to include a additive effect for discrete
time points, \code{FALSE} to omit.}

\item{interaction_baseline}{\code{TRUE} to include baseline-by-time interaction,
\code{FALSE} to omit.}

\item{interaction_group}{\code{TRUE} to include treatment-group-by-time
interaction, \code{FALSE} to omit.}

\item{correlation}{Character of length 1, name of the correlation
structure. Only \code{"unstructured"} is currently supported.}
}
\value{
An object of class \code{"brmsformula"} returned from
\code{brms::brmsformula()}. It contains the fixed effect parameterization,
correlation structure, and residual variance structure.
}
\description{
Build a model formula for an MMRM.
}
\section{Parameterization}{

The formula is not the only factor
that determines the fixed effect parameterization.
The ordering of the categorical variables in the data,
as well as the \code{contrast} option in R, affect the
construction of the model matrix. To see the model
matrix that will ultimately be used in \code{\link[=brm_model]{brm_model()}},
run \code{\link[brms:make_standata]{brms::make_standata()}} and examine the \code{X} element
of the returned list. See the examples below for a
demonstration.
}

\examples{
set.seed(0)
data <- brm_data(
  data = brm_simulate_simple()$data,
  outcome = "response",
  role = "response",
  group = "group",
  time = "time",
  patient = "patient",
  reference_group = "group_1",
  reference_time = "time_1"
)
brm_formula(data)
brm_formula(data = data, intercept = FALSE, effect_baseline = FALSE)
formula <- brm_formula(
  data = data,
  intercept = FALSE,
  effect_baseline = FALSE,
  interaction_group = FALSE
)
formula
# Optional: set the contrast option, which determines the model matrix.
options(contrasts = c(unordered = "contr.SAS", ordered = "contr.poly"))
# See the fixed effect parameterization you get from the data:
head(brms::make_standata(formula = formula, data = data)$X)
# Specify a different contrast method to use an alternative
# parameterization when fitting the model with brm_model():
options(
  contrasts = c(unordered = "contr.treatment", ordered = "contr.poly")
)
# different model matrix than before:
head(brms::make_standata(formula = formula, data = data)$X)
}
\seealso{
Other models: 
\code{\link{brm_model}()}
}
\concept{models}
