% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brm_model.R
\name{brm_model}
\alias{brm_model}
\title{Basic MMRM}
\usage{
brm_model(
  data,
  formula = brm_formula(),
  sd_intercept = 100,
  sd_b = 100,
  sd_sigma = 100,
  shape_cor = 1,
  ...
)
}
\arguments{
\item{data}{A tidy data frame with one row per patient per discrete
time point.}

\item{formula}{An object of class \code{"brmsformula"} from \code{\link[=brm_formula]{brm_formula()}}
or \code{brms::brmsformula()}. Should include the full parameterization
of the model, including fixed effects, residual correlation,
and heterogeneity in the discrete-time-specific residual variance
components.}

\item{sd_intercept}{Positive numeric of length 1, prior standard deviation
of the "intercept" class of parameters.}

\item{sd_b}{Positive numeric of length 1, prior standard deviation
of the "b" class of parameters.}

\item{sd_sigma}{Positive numeric of length 1,
prior standard deviation
of the "b" class of parameters with \code{dpar = "sigma"}.}

\item{shape_cor}{Positive numeric of length 1. For unstructured
correlation, this is the LKJ shape parameter.}

\item{...}{Arguments to \code{brms::brm()} other than \code{data}, \code{formula},
and \code{prior}.}
}
\value{
A fitted model object from \code{brms}.
}
\description{
Fit a basic MMRM model using \code{brms}.
}
\examples{
set.seed(0L)
sim <- brm_simulate()
data <- sim$data
formula <- brm_formula(
  response = "response",
  group = "group",
  time = "time",
  patient = "patient",
  effect_base = FALSE,
  interaction_base = FALSE
)
tmp <- utils::capture.output(
  suppressMessages(
    suppressWarnings(
      model <- brm_model(
        data = data,
        formula = formula,
        chains = 1,
        iter = 100,
        refresh = 0
      )
    )
  )
)
model
brms::prior_summary(model)
}
\seealso{
Other model: 
\code{\link{brm_formula}()}
}
\concept{model}
