% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brm_plot_compare.R
\name{brm_plot_compare}
\alias{brm_plot_compare}
\title{Visually compare the marginals of models and datasets.}
\usage{
brm_plot_compare(
  ...,
  marginal = "response",
  versus = "time",
  facet = c("group", "subgroup")
)
}
\arguments{
\item{...}{Named \code{tibble}s of marginals posterior summaries
from \code{\link[=brm_marginal_summaries]{brm_marginal_summaries()}} and/or \code{\link[=brm_marginal_data]{brm_marginal_data()}}.}

\item{marginal}{Character of length 1, which kind of marginal
to visualize. Must be a value in the \code{marginal} column of the supplied
\code{tibble}s in the \code{...} argument.
Only applies to MCMC output, the data is always on the scale of the
response variable.}

\item{versus}{Character of length 1 identifying the quantity to plot
the marginals against. Must be be one of \code{"time"} or \code{"group"} if
the marginal summaries are not subgroup-specific. If the marginals
are subgroup-specific, then \code{versus} must be one of
\code{"time"}, \code{"group"}, or \code{"subgroup"}.}

\item{facet}{Character vector of length 1 or 2 with quantities to
generate facets. Each element must be \code{"time"}, \code{"group"},
or \code{"subgroup"}, and \code{c(versus, facet)} must all have unique elements.
\code{"subgroup"} is automatically removed if the marginals have no
subgroup. If \code{facet} has length 1, then faceting is wrapped.
If \code{facet} has length 2, then faceting is in a grid,
and the first element is horizontal facet.}
}
\value{
A \code{ggplot} object.
}
\description{
Visually compare the marginals of models and datasets.
}
\examples{
if (identical(Sys.getenv("BRM_EXAMPLES", unset = ""), "true")) {
set.seed(0L)
data <- brm_data(
  data = brm_simulate_simple()$data,
  outcome = "response",
  role = "response",
  group = "group",
  time = "time",
  patient = "patient",
  reference_group = "group_1",
  reference_time = "time_1"
)
formula <- brm_formula(
  data = data,
  baseline = FALSE,
  baseline_time = FALSE
)
tmp <- utils::capture.output(
  suppressMessages(
    suppressWarnings(
      model <- brm_model(
        data = data,
        formula = formula,
        chains = 1,
        iter = 100,
        refresh = 0
      )
    )
  )
)
draws <- brm_marginal_draws(model = model, data = data)
suppressWarnings(summaries_draws <- brm_marginal_summaries(draws))
summaries_data <- brm_marginal_data(data)
brm_plot_compare(
  model1 = summaries_draws,
  model2 = summaries_draws,
  data = summaries_data
)
brm_plot_compare(
  model1 = summaries_draws,
  model2 = summaries_draws,
  marginal = "difference"
)
}
}
\seealso{
Other visualization: 
\code{\link{brm_plot_draws}()}
}
\concept{visualization}
