% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brm_summary.R
\name{brm_summary}
\alias{brm_summary}
\title{Summarize an MMRM.}
\usage{
brm_summary(marginals, level = 0.95, direction = "greater", threshold = 0)
}
\arguments{
\item{marginals}{Posterior draws of the marginal posterior
obtained from \code{\link[=brm_marginals]{brm_marginals()}}.}

\item{level}{Numeric of length 1 between 0 and 1, credible level
for the credible intervals. Only relevant when \code{return = "marginals"}.}

\item{direction}{Character vector of the same length as \code{threshold}.
\code{"greater"} to compute the marginal posterior probability that the
treatment effect is greater than the threshold,
\code{"less"} to compute the marginal posterior probability that the
treatment effect is less than the threshold.
Each element \code{direction[i]} corresponds to \code{threshold[i]}
for all \code{i} from 1 to \code{length(direction)}.}

\item{threshold}{Numeric vector of the same length as \code{direction},
treatment effect threshold for computing posterior probabilities.
Each element \code{direction[i]} corresponds to \code{threshold[i]} for
all \code{i} from 1 to \code{length(direction)}.}
}
\value{
A named list of two data frames with one row per summary
statistic:
\itemize{
\item \code{means}: summary statistics of the marginal posterior mean of
each treatment group and time point. Columns include \code{marginal}
for the type of marginal distribution (see the "Value" section of
the \code{\link[=brm_marginals]{brm_marginals()}} help file for details), \code{group} for treatment
group, \code{time} for discrete time point, \code{statistic} for the type of
summary statistic, \code{value} for the numeric value of the estimate,
and \code{mcse} for the Monte Carlo standard error of the estimate.
\item \code{probabilities}: marginal posterior probabilities of the form
\verb{Prob(treatment effect > threshold | data)} and/or
\verb{Prob(treatment effect < threshold | data)}. Columns include
\code{group} for the treatment group, \code{time} for the discrete time point,
\code{direction} to indicate the direction of the comparison
(\code{"greater"} for \code{>}, \code{"less"} for \code{<}), \code{threshold} for the
treatment effect threshold in the probability statement,
and \code{value} for the numberic value of the estimate of the
probability.
}
}
\description{
Summarize a basic MMRM model fit.
}
\examples{
set.seed(0L)
sim <- brm_simulate()
data <- sim$data
data$group <- paste("treatment", data$group)
data$time <- paste("visit", data$time)
formula <- brm_formula(
  response = "response",
  group = "group",
  time = "time",
  patient = "patient",
  effect_base = FALSE,
  interaction_base = FALSE
)
tmp <- utils::capture.output(
  suppressMessages(
    suppressWarnings(
      model <- brm_model(
        data = data,
        formula = formula,
        chains = 1,
        iter = 100,
        refresh = 0
      )
    )
  )
)
marginals <- brm_marginals(
  model = model,
  group = "group",
  time = "time",
  patient = "patient",
  control = "treatment 1",
  baseline = "visit 1",
  outcome = "response"
)
brm_summary(marginals)
}
\seealso{
Other results: 
\code{\link{brm_marginals}()}
}
\concept{results}
