% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brm_summary.R
\name{brm_summary}
\alias{brm_summary}
\title{Summarize an MMRM.}
\usage{
brm_summary(
  model,
  base = "BASE",
  group = "TRT01P",
  time = "AVISIT",
  patient = "USUBJID",
  covariates = character(0),
  control = "Placebo"
)
}
\arguments{
\item{model}{Fitted \code{brms} model object from \code{\link[=brm_model]{brm_model()}}.}

\item{base}{Character of length 1, name of the baseline response variable
in the data.}

\item{group}{Character of length 1, name of the treatment group
variable in the data.}

\item{time}{Character of length 1, name of the discrete time variable
in the data.}

\item{patient}{Character of length 1, name of the patient ID variable
in the data.}

\item{covariates}{Character vector of names of other covariates
in the data.}

\item{control}{Character of length 1, name of the control arm
in the \code{group} column in the data.}
}
\value{
A \code{tibble} with summary statistics of the marginal posterior.
}
\description{
Summarize a basic MMRM model fit.
}
\details{
Currently assumes the response variable is \code{CHG}
(change from baseline) and not \code{AVAL} (raw response).
}
\examples{
set.seed(0L)
sim <- brm_simulate()
data <- sim$data
formula <- brm_formula(
  response = "response",
  group = "group",
  time = "time",
  patient = "patient",
  effect_base = FALSE,
  interaction_base = FALSE
)
tmp <- utils::capture.output(
  suppressMessages(
    suppressWarnings(
      model <- brm_model(
        data = data,
        formula = formula,
        chains = 1,
        iter = 100,
        refresh = 0
      )
    )
  )
)
brm_summary(
  model = model,
  group = "group",
  time = "time",
  patient = "patient",
  control = 1
)
}
\concept{results}
