% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brm_package.R
\name{brms.mmrm-package}
\alias{brms.mmrm-package}
\title{brms.mmrm: Bayesian MMRMs using \code{brms}}
\description{
The mixed model for repeated measures (MMRM) is a
popular model for longitudinal clinical trial data with
continuous endpoints, and \code{brms} a is powerful and versatile
package for fitting Bayesian regression models.
The \code{brms.mmrm} R package leverages \code{brms} to run MMRMs, and
it supports a simplified interfaced to reduce difficulty
and align with the best practices of the life sciences.
}
\references{
\itemize{
\item Bürkner, P.-C. (2017), "brms: An R package for Bayesian
multilevel models using Stan,"
Journal of Statistical Software, 80, 1–28.
https://doi.org/10.18637/jss.v080.i01.
\item Mallinckrodt, C. H., Lane, P. W., Schnell, D., and others (2008),
"Recommendations for the primary analysis of continuous endpoints
in longitudinal clinical trials,"
Ther Innov Regul Sci, 42, 303–319.
https://doi.org/10.1177/009286150804200402.
\item Mallinckrodt, C. H., and Lipkovich, I. (2017),
Analyzing longitudinal clinical trial data: A practical guide,
CRC Press, Taylor & Francis Group.
}
}
\concept{help}
