% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdystonia.R
\docType{data}
\name{cdystonia}
\alias{cdystonia}
\title{cdystonia dataset}
\format{
A tibble with 420 rows and 6 variables

\describe{
\item id. ...
\item sex. ...
\item age. ...
\item baseline. ...
\item treat, ...
\item week. ...
\item chg ...
}
}
\usage{
data(cdystonia)
}
\description{
Subset from a multicenter, randomized controlled trial of botulinum toxin
type B in patients with cervical dystonia.
}
\references{
\url{https://hbiostat.org/data/repo/cdystonia}
}
\keyword{datasets}
