% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdystonia.R
\docType{data}
\name{cdystonia}
\alias{cdystonia}
\title{cdystonia dataset}
\format{
A tibble with 347 rows and 7 variables

\describe{
\item{id}{id}
\item{sex}{sex}
\item{age}{age}
\item{baseline}{baseline}
\item{treat}{treat}
\item{week}{week}
\item{chg}{chg}
}
}
\usage{
cdystonia
}
\description{
Subset from a multicenter, randomized controlled trial of botulinum toxin
type B in patients with cervical dystonia.
}
\keyword{datasets}
